+++ src/Collection/Collection.php
@@ -71,7 +71,7 @@ class Collection extends ArrayObject implements JsonSerializable
     /**
      * @inheritDoc
      */
-    public function jsonSerialize()
+    public function jsonSerialize(): array
     {
         return $this->getArrayCopy();
     }