+++ CHANGELOG.md
@@ -20,6 +20,7 @@ use [Keep a Changelog] (http://keepachangelog.com/).
 - @package attributes from PhpDoc
 - Unused `\Hector\Orm\ExternalEnvironment` class
 - Unused `\Hector\Orm\OrmTrait` class
+- Composer JSON extension requirement, already included in PHP 8
 
 ## [1.0.0-beta1] - 2021-06-02