+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file. This projec
 to [Semantic Versioning] (http://semver.org/). For change log format,
 use [Keep a Changelog] (http://keepachangelog.com/).
 
+## [1.0.0-beta3] - In progress
+
+### Fixed
+
+- Signature of `Collection::jsonSerialize(): array`
+- Signature of `MagicEntity::jsonSerialize(): array`
+
 ## [1.0.0-beta2] - 2021-07-07
 
 ### Added